email_assistant/requirements.txt
burr[start,opentelemetry]
gradio
pymupdf
opentelemetry-instrumentation-openai
opentelemetry-instrumentation-ollama
# instructor # note this has some conflicts with llama index.
# jupyter # required for notebook to run